Completed POC and uploaded 1st workable build
This is a fully working proof of concept of the log assistant tool. It is now able to take a thread log and parse through and display specific information. 

Tested ability to package it into a runable self executing jar file.
Tested ability to read and correctly parse logs from CS 10.5 and 16.2.